+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## v1.0.1 - Unreleased
 ### Changed
 - Change hint text "It's the GUY with ..." to be non-gendered. It can be assumed that we are referring to men only.
+- Change score/points to money "$" to be more realistic.
 
 ### Fixed
 - Increase game over text Z-index